Moved Datafile methods to DatafileMixin
