+++ pepys_import/file/gpx_importer.py
@@ -105,7 +105,9 @@ class GPXImporter(Importer):
 
                 # Parse timestamp and create state
                 timestamp = parse(timestamp_str)
-                state = datafile.create_state(sensor, timestamp, self.short_name)
+                state = datafile.create_state(
+                    data_store, sensor, timestamp, self.short_name
+                )
 
                 # Add location (no need to convert as it requires a string)
                 if track_name in prev_location:
